ProcessInitTime type to provide compat with earlier versions of runc (<=v1.0.0-rc3)

+++ libcontainer/factory_linux.go
@@ -200,7 +200,7 @@ func (l *LinuxFactory) Load(id string) (Container, error) {
 	}
 	r := &nonChildProcess{
 		processPid:       state.InitProcessPid,
-		processStartTime: state.InitProcessStartTime,
+		processStartTime: uint64(state.InitProcessStartTime),
 		fds:              state.ExternalDescriptors,
 	}
 	// We have to use the RootlessManager.
@@ -209,7 +209,7 @@ func (l *LinuxFactory) Load(id string) (Container, error) {
 	}
 	c := &linuxContainer{
 		initProcess:          r,
-		initProcessStartTime: state.InitProcessStartTime,
+		initProcessStartTime: uint64(state.InitProcessStartTime),
 		id:                   id,
 		config:               &state.Config,
 		initArgs:             l.InitArgs,
